website/requirements.txt
appdirs==1.4.4
bleach==3.2.1
dj-static==0.0.6
Django==2.2
django-filter==2.3.0
django-session-timeout==0.1.0
djangorestframework==3.11.1
gunicorn==20.0.4
Pillow==7.2.0
static3==0.7.0
virtualenv==20.0.33
